requirements.txt
appdirs==1.4.4
argcomplete==3.6.3
guizero==1.6.0
importlib_resources==6.5.2
pillow==12.1.1
python-barcode==0.16.1
python-escpos==3.1
pyusb==1.3.1
PyYAML==6.0.3
qrcode==8.2
setuptools==82.0.1
six==1.17.0
usb==0.0.83.dev0
